Check whether Input is Valid

Checks whether the input for the function CVN is valid. This function does not return anything. The execution of the function halts when an issue has been detected.

Usage

check_correctness_input(raw_data, W, lambda1, lambda2, gamma1, gamma2, rho)

Arguments

raw_data

A list with matrices. The number of columns should be the same for each matrix

W

The \((m \times m)\)-dimensional upper-triangular weight matrix \(W\)

lambda1

A vector of \(\lambda_1\)'s LASSO penalty terms

lambda2

A vector of \(\lambda_2\)'s global smoothing parameters

gamma1

A vector of \(\gamma_1\)'s LASSO penalty terms. Note that \(\gamma_1 = \frac{2 \lambda_1}{m p (1 - p)}\)

gamma2

A vector of \(\gamma_2\)'s global smoothing parameters. Note that \(\gamma_2 = \frac{4 \lambda_2}{m(m-1)p(p-1)}\)

rho

The \(\rho\) ADMM's penalty parameter

See also