Different Weight Matrices

This function generates different weight matrices for two external covariates, \(W\). There are several types:

full

All graphs are fully connected with weight 1

glasso

All graphs are disconnected with weight 0. This mimicks the GLASSO, where each graph is estimated independently

grid

A weight matrix for a \(k \times l\) grid

uniform-random

Fully-connected, but the entries are drawn from a uniform distribution

Usage

create_weight_matrix(
  type = c("full", "glasso", "grid", "uniform-random"),
  k,
  l,
  plot = FALSE
)

Arguments

type

The type of weight matrix

k

Number of categories in the first external covariate

l

Number of categories in the second external covariate

plot

If TRUE, the weight matrix is plotted (default: FALSE)

Value

Weight matrix

Examples

W <- create_weight_matrix(type="grid", k=3, l=2)