Edge Overview

Function that gives an overview about the number of edges in each subgraph over all fitted CVN models.

Usage

cvn_edge_summary(cvn)

Arguments

cvn

A cvn object

Value

A data.frame.Shows the number of edges in each subgraph, the number of edges present in all graphs (core edges) and the number of edges that are unique for a subgraph

Examples

data(grid)

W <- create_weight_matrix("grid", 3, 3)

# lambdas:
lambda1 = 1  # can also be lambda1 = 1:2 
lambda2 = .2

fit <- CVN(grid, W, lambda1 = lambda1, lambda2 = lambda2)
#> Estimating a CVN with 9 graphs...
#> 
#> Number of cores: 1
#> Uses a warmstart...
#> 
#> -------------------------
#> iteration 1  |  2.180956
#> iteration 2  |  0.115992
#> iteration 3  |  0.074886
#> iteration 4  |  0.048989
#> iteration 5  |  0.026376
#> iteration 6  |  0.009400
#> iteration 7  |  0.006532
#> iteration 8  |  0.004188
#> iteration 9  |  0.003427
#> iteration 10  |  0.002602
#> -------------------------
#> iteration 11  |  0.002151
#> iteration 12  |  0.002050
#> iteration 13  |  0.001641
#> iteration 14  |  0.001472
#> iteration 15  |  0.001493
#> iteration 16  |  0.001085
#> iteration 17  |  0.000864
#> iteration 18  |  0.000816
#> iteration 19  |  0.000768
#> iteration 20  |  0.000608
#> -------------------------
#> iteration 21  |  0.000582
#> iteration 22  |  0.000563
#> iteration 23  |  0.000575
#> iteration 24  |  0.000516
#> iteration 25  |  0.000486
#> iteration 26  |  0.000449
#> iteration 27  |  0.000419
#> iteration 28  |  0.000404
#> iteration 29  |  0.000372
#> iteration 30  |  0.000355
#> -------------------------
#> iteration 31  |  0.000335
#> iteration 32  |  0.000322
#> iteration 33  |  0.000300
#> iteration 34  |  0.000281
#> iteration 35  |  0.000272
#> iteration 36  |  0.000262
#> iteration 37  |  0.000257
#> iteration 38  |  0.000245
#> iteration 39  |  0.000236
#> iteration 40  |  0.000227
#> -------------------------
#> iteration 41  |  0.000213
#> iteration 42  |  0.000205
#> iteration 43  |  0.000193
#> iteration 44  |  0.000183
#> iteration 45  |  0.000175
#> iteration 46  |  0.000169
#> iteration 47  |  0.000163
#> iteration 48  |  0.000157
#> iteration 49  |  0.000152
#> iteration 50  |  0.000140
#> -------------------------
#> iteration 51  |  0.000136
#> iteration 52  |  0.000109
#> iteration 53  |  0.000114
#> iteration 54  |  0.000102
#> iteration 55  |  0.000101
#> iteration 56  |  0.000093
                 
cvn_edge_summary(fit)
#>   edges core_edges unique_edges
#> 1    84         34            0
#> 2    86         34            0
#> 3    84         34            0
#> 4    82         34            0
#> 5    84         34            0
#> 6    88         34            0
#> 7    88         34            0
#> 8    90         34            0
#> 9    84         34            0