Determine matrix \(A\) for inner-ADMM for the \(Z\)-update step

The \(Z\)-update step requires us to solve a special Generalized LASSO problem of the form $$ \hat{\beta} = \text{argmin } \frac{1}{2} || y - \beta ||_2^2 + ||D\beta||_1 $$ where \(\beta\) and \(y\) are \(m\)-dimensional vectors and \(D\) is a \((c \times m)\)-matrix where \(c = (m^2 + m) / 2\). We solve this optimization problem using an adaption of the ADMM algorithm presented in Zhu (2017). This algorithm requires the choice of a matrix \(A\) such that \(A - D'D\) is positive semidefinite. In order to optimize the ADMM, we choose the matrix \(A\) to be diagonal with a fixed value \(a\). This function determines the smallest value of \(a\) such that \(A - D'D\) is indeed positive semidefinite. We do this be determining the largest eigenvalue

Usage

matrix_A_inner_ADMM(W, eta1, eta2)

Arguments

W

Weight matrix \(W\)

eta1, eta2

The values \(\eta_1 = \lambda_1 / \rho\) and \(\eta_2 = \lambda_2 / \rho\)

Value

Value of \(a\)

References

Zhu, Y. (2017). An Augmented ADMM Algorithm With Application to the Generalized Lasso Problem. Journal of Computational and Graphical Statistics, 26(1), 195–204.
doi:10.1080/10618600.2015.1114491

Author

Louis Dijkstra