All visNetwork plots for a CVN object

Creates all visNetwork plots, see visnetwork, for all graphs in a cvn object

Usage

visnetwork_cvn(
  cvn,
  node_titles = 1:cvn$p,
  titles = lapply(1:cvn$n_lambda_values, function(i) sapply(1:cvn$m, function(j) "")),
  show_core_graph = TRUE,
  width = c(3, 1),
  color = c("red", "blue"),
  igraph_layout = "layout_in_circle",
  verbose = TRUE
)

Arguments

cvn

A cvn object, see CVN

node_titles

Vector with title of the nodes (Default: 1:p)

titles

A list with n_lambda_values vectors. Each vector is of the lenght m. Regulates the titles of the graphs (Default: no title)

show_core_graph, width, color

Show the core graph using the width and colors

igraph_layout

igraph layout (default: layout_in_circle)

verbose

Verbose (Default: TRUE)

Value

List

See also

Examples

path <- system.file("cvnfit.RData", package = "CVN")
load(path)