Edge Overview

Function that gives an overview about the number of edges in each subgraph over all fitted CVN models.

Usage

cvn_edge_summary(cvn)

Arguments

cvn

A cvn object

Value

A data.frame showing the number of edges in each subgraph, the number of edges present in all graphs (core edges) and the number of edges that are unique for a subgraph

Examples

data(grid)
W <- create_weight_matrix("grid", 3, 3)

# lambdas:
lambda1 = c(1, 1.5)
lambda2 = .2

fit <- CVN(grid, W, lambda1 = lambda1, lambda2 = lambda2, n_cores = 1)
#> Estimating a CVN with 9 graphs...
#> 
#> Number of cores: 1
#> Uses a warmstart...
#> 
#> -------------------------
#> iteration 1  |  2.180956
#> iteration 2  |  0.115992
#> iteration 3  |  0.074886
#> iteration 4  |  0.048989
#> iteration 5  |  0.026376
#> iteration 6  |  0.009400
#> iteration 7  |  0.006532
#> iteration 8  |  0.004188
#> iteration 9  |  0.003427
#> iteration 10  |  0.002602
#> -------------------------
#> iteration 11  |  0.002151
#> iteration 12  |  0.002050
#> iteration 13  |  0.001641
#> iteration 14  |  0.001472
#> iteration 15  |  0.001493
#> iteration 16  |  0.001085
#> iteration 17  |  0.000864
#> iteration 18  |  0.000816
#> iteration 19  |  0.000768
#> iteration 20  |  0.000608
#> -------------------------
#> iteration 21  |  0.000582
#> iteration 22  |  0.000563
#> iteration 23  |  0.000575
#> iteration 24  |  0.000516
#> iteration 25  |  0.000486
#> iteration 26  |  0.000449
#> iteration 27  |  0.000419
#> iteration 28  |  0.000404
#> iteration 29  |  0.000372
#> iteration 30  |  0.000355
#> -------------------------
#> iteration 31  |  0.000335
#> iteration 32  |  0.000322
#> iteration 33  |  0.000300
#> iteration 34  |  0.000281
#> iteration 35  |  0.000272
#> iteration 36  |  0.000262
#> iteration 37  |  0.000257
#> iteration 38  |  0.000245
#> iteration 39  |  0.000236
#> iteration 40  |  0.000227
#> -------------------------
#> iteration 41  |  0.000213
#> iteration 42  |  0.000205
#> iteration 43  |  0.000193
#> iteration 44  |  0.000183
#> iteration 45  |  0.000175
#> iteration 46  |  0.000169
#> iteration 47  |  0.000163
#> iteration 48  |  0.000157
#> iteration 49  |  0.000152
#> iteration 50  |  0.000140
#> -------------------------
#> iteration 51  |  0.000136
#> iteration 52  |  0.000109
#> iteration 53  |  0.000114
#> iteration 54  |  0.000102
#> iteration 55  |  0.000101
#> iteration 56  |  0.000093
#> -------------------------
#> iteration 1  |  1.887193
#> iteration 2  |  0.135032
#> iteration 3  |  0.091797
#> iteration 4  |  0.051830
#> iteration 5  |  0.029767
#> iteration 6  |  0.023468
#> iteration 7  |  0.011319
#> iteration 8  |  0.007583
#> iteration 9  |  0.006705
#> iteration 10  |  0.004910
#> -------------------------
#> iteration 11  |  0.004122
#> iteration 12  |  0.003544
#> iteration 13  |  0.003412
#> iteration 14  |  0.002673
#> iteration 15  |  0.002175
#> iteration 16  |  0.001636
#> iteration 17  |  0.001523
#> iteration 18  |  0.001362
#> iteration 19  |  0.001236
#> iteration 20  |  0.001114
#> -------------------------
#> iteration 21  |  0.001057
#> iteration 22  |  0.001067
#> iteration 23  |  0.000890
#> iteration 24  |  0.000790
#> iteration 25  |  0.000728
#> iteration 26  |  0.000669
#> iteration 27  |  0.000636
#> iteration 28  |  0.000607
#> iteration 29  |  0.000556
#> iteration 30  |  0.000526
#> -------------------------
#> iteration 31  |  0.000482
#> iteration 32  |  0.000464
#> iteration 33  |  0.000436
#> iteration 34  |  0.000424
#> iteration 35  |  0.000383
#> iteration 36  |  0.000331
#> iteration 37  |  0.000316
#> iteration 38  |  0.000308
#> iteration 39  |  0.000319
#> iteration 40  |  0.000290
#> -------------------------
#> iteration 41  |  0.000277
#> iteration 42  |  0.000259
#> iteration 43  |  0.000251
#> iteration 44  |  0.000218
#> iteration 45  |  0.000194
#> iteration 46  |  0.000216
#> iteration 47  |  0.000192
#> iteration 48  |  0.000183
#> iteration 49  |  0.000177
#> iteration 50  |  0.000171
#> -------------------------
#> iteration 51  |  0.000165
#> iteration 52  |  0.000160
#> iteration 53  |  0.000151
#> iteration 54  |  0.000130
#> iteration 55  |  0.000115
#> iteration 56  |  0.000106
#> iteration 57  |  0.000103
#> iteration 58  |  0.000100

# Edge summary for a list of CVN models
cvn_edge_summary(fit)
#>   E(g1) E(g2) E(g3) E(g4) E(g5) E(g6) E(g7) E(g8) E(g9) E(core) E(g1_u) E(g2_u)
#> 1    42    43    42    41    42    44    44    45    42      34       0       0
#> 2    43    42    43    41    40    41    43    45    41      32       0       0
#>   E(g3_u) E(g4_u) E(g5_u) E(g6_u) E(g7_u) E(g8_u) E(g9_u)
#> 1       0       0       0       0       0       0       0
#> 2       0       0       0       0       0       0       0

# Edge summary for a single CVN
fit2 <- extract_cvn(fit, id = 2)
cvn_edge_summary(fit2)                 
#>   edges core_edges unique_edges
#> 1    43         32            0
#> 2    42         32            0
#> 3    43         32            0
#> 4    41         32            0
#> 5    40         32            0
#> 6    41         32            0
#> 7    43         32            0
#> 8    45         32            0
#> 9    41         32            0